pyproject.toml
[tool.poetry]
name = "npc"
version = "0.1.0"
description = ""
authors = ["max brewer <58602708+deepfates@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.10"
langchain = "^0.0.59"
textworld = "^1.5.3"
python-dotenv = "^0.21.0"
openai = "^0.26.0"
tiktoken = "^0.1.2"
Flask = {extras = ["async"], version = "^2.2.2"}
replicate = "^0.4.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.20.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
